requirements.txt
openai
python-dotenv
anthropic
aiogram
